Fix tests

+++ lang/java/avro/src/main/java/org/apache/avro/generic/PrimitivesArrays.java
@@ -97,7 +97,7 @@ public class PrimitivesArrays {
 
     /**
      * Direct primitive int access.
-     * 
+     *
      * @param i : index.
      * @return value at index.
      */
@@ -204,7 +204,7 @@ public class PrimitivesArrays {
 
     /**
      * Direct primitive int access.
-     * 
+     *
      * @param i : index.
      * @return value at index.
      */
@@ -318,7 +318,7 @@ public class PrimitivesArrays {
 
     /**
      * Direct primitive int access.
-     * 
+     *
      * @param i : index.
      * @return value at index.
      */
@@ -479,7 +479,7 @@ public class PrimitivesArrays {
 
     /**
      * Direct primitive int access.
-     * 
+     *
      * @param i : index.
      * @return value at index.
      */
@@ -586,7 +586,7 @@ public class PrimitivesArrays {
 
     /**
      * Direct primitive int access.
-     * 
+     *
      * @param i : index.
      * @return value at index.
      */
@@ -601,7 +601,7 @@ public class PrimitivesArrays {
       if (o == null) {
         return;
       }
-      this.add(location, o.floatValue());
+      this.add(location, o.doubleValue());
     }
 
     public void add(int location, double o) {
